+++ ChangeLog.md
@@ -8,6 +8,8 @@
   ([#373](https://github.com/davep/rogallo/pull/373))
 - Added the ability to export certificates.
   ([#389](https://github.com/davep/rogallo/pull/389))
+- Added the ability to import certificates.
+  ([#389](https://github.com/davep/rogallo/pull/389))
 - Removed the `ToggleHistoryManager` command.
   ([#373](https://github.com/davep/rogallo/pull/373))
 - Removed the `ToggleBookmarksManager` command.